requirements.txt
beautifulsoup4
gspread
oauth2client
python-dotenv
requests
